Implemented delete save functions to the managers
It solved the problem that caused the last session rent button to fire the click event in the new session, renting two devices at once, This problem occurred because when we left a savegame and starts another, the button keeps the listeners on the onclick event.
So a solution , not so clean, when session is closed the button clears all events listeners , The most correct solution would be to remove the button from modEventListeners and erase the object from the memory.

- This closes #3